Separation of concern for calendar modal
