+++ src/components/calendar/modules/CalendarModal.js
@@ -3,11 +3,7 @@ import Calendar from './Calendar';
 import styled from 'styled-components';
 import { colors } from '../../../../utils/constants/colors';
 import { Button, Heading, Stack } from '@kiwicom/orbit-components/lib';
-import { getFormattedDateRange } from '../../../../utils/api/time';
 import Modal, { ModalSection, ModalFooter } from '@kiwicom/orbit-components/lib/Modal';
-import { donations } from '../../../../utils/constants/postType';
-import { useRouter } from 'next/router';
-import api from '../../../../utils/api';
 
 const okButton = styled.button`
   background: ${colors.calendarSelectedBackground};
@@ -27,47 +23,13 @@ const DATE = '5';
 const title = 'Suggest Dates to deliver';
 const description = '*Please select up to ' + DATE + ' dates';
 
-const CalendarModal = ({
-  postType,
-  postId,
-  selectedChatId,
-  setSelectedChatId,
-  isNewChat,
-  setIsNewChat,
-  onShow,
-  onHide,
-}) => {
+const CalendarModal = ({ onShow, onHide, sendCalendarMessage }) => {
   const [selectedDates, setSelectedDates] = useState([]);
-  const router = useRouter();
   if (!onShow) {
     return <div></div>;
   }
   const onClickOk = () => {
-    if (selectedDates.length > 0) {
-      const message = selectedDates
-        .map((selectedDate) => getFormattedDateRange(selectedDate.startDate, selectedDate.endDate))
-        .join(','); // dates are separated by comma
-      if (isNewChat) {
-        sendFirstCalendarMessage(message)
-          .then((chat) => {
-            // need to get the chat id from the newly created chat to select chat id
-            setSelectedChatId(chat.chatId);
-            setIsNewChat(false);
-          })
-          .catch((err) => console.error(err));
-      } else {
-        api.chats.sendCalendarMessage(selectedChatId, message).catch((err) => console.error(err));
-      }
-    }
-    onHide();
-  };
-
-  const sendFirstCalendarMessage = async (calendarRawString) => {
-    const method =
-      postType === donations ? 'sendInitialCalendarMessageForDonation' : 'sendInitialCalendarMessageForWish';
-    const [rawChat, rawFirstMessage] = await api.chats[method](postId, calendarRawString);
-    router.push(`/chat/${rawChat.data().chatId}`);
-    return rawChat.data();
+    sendCalendarMessage(selectedDates);
   };
 
   return (
