+++ src/components/chat/modules/ChatDialogUserRow.js
@@ -13,6 +13,10 @@ import { colors } from '../../../../utils/constants/colors';
 import { CardSection } from '@kiwicom/orbit-components/lib/Card';
 import { PENDING, COMPLETED } from '../../../../utils/constants/postStatus';
 import StatusTag from '../../../components/tags/StatusTag';
+import { donations } from '../../../../utils/constants/postType';
+import { useRouter } from 'next/router';
+import api from '../../../../utils/api';
+import { getFormattedDateRange } from '../../../../utils/api/time';
 
 const AvatarContainer = styled.div`
   float: left;
@@ -47,6 +51,7 @@ const ChatDialogUserRow = ({
   const [showAppreciationMessageModal, setShowAppreciationMessageModal] = useState(false);
   const [showConfirmCompletionModal, setShowConfirmCompletionModal] = useState(false);
   const [status, setStatus] = useState(postStatus);
+  const router = useRouter();
 
   const handleCloseAppreciationMessageModal = () => setShowAppreciationMessageModal(false);
   const handleShowAppreciationMessageModal = () => setShowAppreciationMessageModal(true);
@@ -64,6 +69,34 @@ const ChatDialogUserRow = ({
   useEffect(() => {
     setStatus(postStatus);
   }, [postStatus]);
+
+  const handleSendCalendarMessage = (selectedDates) => {
+    if (selectedDates.length > 0) {
+      const message = selectedDates
+        .map((selectedDate) => getFormattedDateRange(selectedDate.startDate, selectedDate.endDate))
+        .join(','); // dates are separated by comma
+      if (isNewChat) {
+        sendFirstCalendarMessage(message)
+          .then((chat) => {
+            // need to get the chat id from the newly created chat to select chat id
+            setSelectedChatId(chat.chatId);
+            setIsNewChat(false);
+          })
+          .catch((err) => console.error(err));
+      } else {
+        api.chats.sendCalendarMessage(selectedChatId, message).catch((err) => console.error(err));
+      }
+    }
+    handleCloseSuggestDateModal();
+  };
+
+  const sendFirstCalendarMessage = async (calendarRawString) => {
+    const method =
+      postType === donations ? 'sendInitialCalendarMessageForDonation' : 'sendInitialCalendarMessageForWish';
+    const [rawChat, rawFirstMessage] = await api.chats[method](postId, calendarRawString);
+    router.push(`/chat/${rawChat.data().chatId}`);
+    return rawChat.data();
+  };
 
   return (
     <CardSection>
@@ -83,14 +116,9 @@ const ChatDialogUserRow = ({
               Suggest Dates
             </Button>
             <CalendarModal
-              postType={postType}
-              postId={postId}
-              selectedChatId={selectedChatId}
-              setSelectedChatId={setSelectedChatId}
-              isNewChat={isNewChat}
-              setIsNewChat={setIsNewChat}
               onShow={showSuggestDateModal}
               onHide={handleCloseSuggestDateModal}
+              sendCalendarMessage={handleSendCalendarMessage}
             />
             {status === PENDING ? (
               <Button
